// Project/InternalAuditSystem.Web/ClientApp/src/app/swaggerapi/AngularFiles/model/aCMPresentationAC.ts
/**
 * My Title
 * No description provided (generated by Swagger Codegen https://github.com/swagger-api/swagger-codegen)
 *
 * OpenAPI spec version: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */
import { ACMDocumentAC } from './aCMDocumentAC';
import { ACMReportMappingAC } from './aCMReportMappingAC';
import { ACMReportStatus } from './aCMReportStatus';
import { ACMStatus } from './aCMStatus';
import { ACMTableAC } from './aCMTableAC';
import { BaseModelAC } from './baseModelAC';
import { EntityUserMappingAC } from './entityUserMappingAC';
import { KeyValuePairOfIntegerAndString } from './keyValuePairOfIntegerAndString';
import { RatingAC } from './ratingAC';
import { ReportStage } from './reportStage';
import { UserAC } from './userAC';

export interface ACMPresentationAC extends BaseModelAC { 
    reportId?: string;
    heading?: string;
    recommendation?: string;
    observation?: string;
    managementResponse?: string;
    ratingId?: string;
    ratingsList?: Array<RatingAC>;
    rating?: RatingAC;
    ratings?: string;
    implication?: string;
    status: ACMStatus;
    statusString?: string;
    acmReportTitle?: string;
    acmReportStatus: ACMReportStatus;
    acmReportStage: ReportStage;
    acmReportStatusString?: string;
    entityId?: string;
    isDraft: boolean;
    isDraftToString?: string;
    acmFiles?: Array<Blob>;
    acmDocuments?: Array<ACMDocumentAC>;
    acmTableList?: Array<ACMTableAC>;
    personResponsible: string;
    personResponsibleList?: Array<UserAC>;
    userList?: Array<EntityUserMappingAC>;
    reviewerList?: Array<ACMReportMappingAC>;
    reviewerStatus?: Array<KeyValuePairOfIntegerAndString>;
}